Deleting semicolons, refactoring the filterbox file, adding hooks
