+++ web-page/src/components/FilterButton.js
@@ -1,15 +1,14 @@
-import React from 'react';
+import React from 'react'
 
-import './FilterButton.css';
+import './FilterButton.css'
 
 const FilterButton = (props) => {
-    const { id, fieldname, customClass } = props;
+    const { id, fieldname, customClass, onClick } = props
     return <label
                 id={id} 
                 className={`${customClass ? 'filter-on' : ''} btn btn-secondary`}
-                fieldname={fieldname}
-                onClick={props.onClick}
-            >{id}</label>;
+                onClick={ _ => onClick(id, fieldname) }
+            >{id}</label>
 }
 
 export default FilterButton;
