+++ web-page/src/components/filter-box/index.js
@@ -1,0 +1,215 @@
+import React, { useState } from 'react'
+import { connect } from 'react-redux'
+import { searchCards, addFilter, removeFilter } from '../../redux/modules/search'
+import './index.css'
+import Filter from '../Filter'
+import { parseFilterText, findArrayItemsInArrayOrString, getFieldsToSearch, mapIdToColor, createAllButtons } from './filter.utils'
+
+const FilterBox = ({ totalCards, appliedFilters, fieldOptions, searchCards, onFilterAdd, onFilterRemove }) => {
+    const [ filterValues, setFilter ] = useState({ filterText: '', fieldToSearch: '', isFilterNegation: false })
+    const { filterText, isFilterNegation } = filterValues
+
+    const onAddFilterButtonHandler = (id, fieldName) => {
+        const filterText = fieldName === "color" ? mapIdToColor[id] : id.toString(), type = 'string'
+        let isFilterAdd = appliedFilters.find(a=> a.id.toLocaleLowerCase() === `${fieldName}:  ${filterText}`.toLocaleLowerCase())
+        if (isFilterAdd) {
+            onFilterRemove(isFilterAdd.id)
+            return// Filter already added
+        }
+        onAddFilter(type, fieldName, filterText, isFilterNegation)
+    }
+
+    const onAddFilterClickHandler = _ => {
+        const { filterText: origText, fieldToSearch: { type, fieldName }, isFilterNegation } = filterValues
+        let filterText = origText.trim()
+        if (appliedFilters.find(a => a.id === `${fieldName}-${filterText}`)) {
+            return// Filter already added
+        }
+        onAddFilter(type, fieldName, filterText, isFilterNegation)
+    }
+
+    const onAddFilter = (type, fieldName, filterText, isFilterNegation) => {
+        if (typeof onFilterAdd !== 'function') {
+            return// No handler, so do nothing
+        }
+
+        if (typeof filterText !== 'string') {
+            return// Empty string
+        }
+
+        const filterConditions = parseFilterText(filterText)
+        let filter = card => {
+            let criteriaToSearchOn
+            if (type === 'object') {// no implementation yet, so we search in the whole object
+                criteriaToSearchOn = [JSON.stringify( card[fieldName] )]
+            }
+            else {
+                let val = card[fieldName]
+                if (val === null || val === undefined) {
+                    return false
+                }
+                criteriaToSearchOn = Array.isArray(val) ? val.slice() : [val]
+
+                // We look for the field in the back of the card and add it to our search
+                if (card['cardBack'] && card['cardBack'][fieldName]) {
+                    val = card['cardBack'][fieldName]
+                    if (Array.isArray(val)) {
+                        criteriaToSearchOn = criteriaToSearchOn.concat(val)
+                    }
+                    else {
+                        criteriaToSearchOn.push( val )
+                    }
+                }
+            }
+
+            return findArrayItemsInArrayOrString(filterConditions, criteriaToSearchOn)
+        }
+
+        if (isFilterNegation) {
+            const oldFilter = filter
+            filter = card => !oldFilter(card)
+        }
+        onFilterAdd(`${fieldName}: ${isFilterNegation ? 'NOT ' : ''} ${filterText}`, filter)
+    }
+
+    const onFieldSelectionChangeHandler = event => {
+        setFilter({ ...filterValues, fieldToSearch: fieldOptions[event.target.value] })
+    }
+
+    const onFilterTextChangeHandler = (event) => {
+        setFilter({ ...filterValues, filterText: event.target.value })
+    }
+
+    const onNegationFilterChangeHandler = (event) => {
+        setFilter({ ...filterValues, isFilterNegation: event.target.checked })
+    }
+
+    const onInputTextKeyDownHandler = (event) => {
+        const isEnter = event.keyCode === 13
+        if (isEnter) {
+            onAddFilterClickHandler()
+        }
+    }
+
+    const removedOptions = ['availableDate', 'cardImageUrl', 'cardBack', 'era', /*'type', 'color', 'energy', 'comboEnergy', 'rarity', 'character', 'skillKeywords', 'cardNumber', */]
+    const optionsToSelect = fieldOptions.map(
+        (option, index) =>
+            !removedOptions.includes(option.fieldName)
+                ? <option key={option.fieldName} value={index}>
+                    {option.label}
+                </option>
+                : null
+    )
+
+    const filtersApplied = appliedFilters.map(
+        ({ id }) => <Filter key={id} id={id} onFilterRemove={onFilterRemove} />
+    )
+
+    const allButtons = createAllButtons(appliedFilters, onAddFilterButtonHandler)
+
+    return (
+        <div className="col s12 filter-box white-text">
+            <div className="row">
+                <div className="input-field col s12">
+                    <div htmlFor="type">Card Type</div>
+                    <div className="mx-auto btn-group-toggle btn-group">
+                        {allButtons.type}
+                    </div>
+                </div>
+                <div className="input-field col s12">
+                    <div htmlFor="color">Color</div>
+                    <div id="color" className="mx-auto btn-group-toggle btn-group">
+                        {allButtons.color}
+                    </div>
+                </div>
+                <div className="input-field col s12">
+                    <div htmlFor="energy">Energy</div>
+                    <div id="energy" className="mx-auto btn-group-toggle btn-group">
+                        {allButtons.energy}
+                    </div>
+                </div>
+                <div className="input-field col s12">
+                    <div htmlFor="cboenergy">Combo Energy</div>
+                    <div id="cboenergy" className="mx-auto btn-group-toggle btn-group">
+                        {allButtons.comboEnergy}
+                    </div>
+                </div>
+                <div className="input-field col s12">
+                    <select id="rarity" className="mx-auto btn-group-toggle btn-group">
+                        <option value="">Rarity</option>
+                        <option value="Common[C]">Common[C]</option>
+                        <option value="Rare[R]">Rare[R]</option>
+                        <option value="Starter Rare[ST]">Starter Rare[ST]</option>
+                        <option value="Super Rare[SR]">Super Rare[SR]</option>
+                        <option value="Uncommon[UC]">Uncommon[UC]</option>
+                        <option value="Other">Other</option>
+                    </select>
+                </div>
+                <div className="input-field col s12">
+                    <select id="character" className="mx-auto btn-group-toggle btn-group">
+                        <option value="">Character</option>
+                        <option value="Common[C]">Common[C]</option>
+                        <option value="Rare[R]">Rare[R]</option>
+                        <option value="Starter Rare[ST]">Starter Rare[ST]</option>
+                        <option value="Super Rare[SR]">Super Rare[SR]</option>
+                        <option value="Uncommon[UC]">Uncommon[UC]</option>
+                        <option value="Other">Other</option>
+                    </select>
+                </div>
+                <div className="input-field col s12">
+                    <select id="skillKeywords" className="mx-auto btn-group-toggle btn-group">
+                        <option value="">Skill Keywords</option>
+                        <option value="Common[C]">Common[C]</option>
+                        <option value="Rare[R]">Rare[R]</option>
+                        <option value="Starter Rare[ST]">Starter Rare[ST]</option>
+                        <option value="Super Rare[SR]">Super Rare[SR]</option>
+                        <option value="Uncommon[UC]">Uncommon[UC]</option>
+                        <option value="Other">Other</option>
+                    </select>
+                </div>
+            </div>
+            <div className="input-field col s12">
+                <select id="filter-box-type-input" defaultValue={0} onChange={onFieldSelectionChangeHandler}> {optionsToSelect} </select>
+                <label htmlFor="filter-box-type-input">Field</label>
+            </div>
+
+            <div className="input-field col s12">
+                <input
+                    id="filter-box-criteria-input"
+                    type="text" placeholder="For an OR operation use ||"
+                    value={filterText}
+                    onChange={onFilterTextChangeHandler}
+                    onKeyDown={onInputTextKeyDownHandler} />
+                <label htmlFor="filter-box-criteria-input">Search</label>
+            </div>
+
+            <div className="input-field col s3">
+                <label>
+                    <input type="checkbox" checked={isFilterNegation} onChange={onNegationFilterChangeHandler}/>
+                    <span>
+                        { isFilterNegation ? 'Not' : null }
+                    </span>
+                </label>
+            </div>
+
+            <div className="input-field col s3">
+                <span><button className="waves-effect waves-light btn" onClick={onAddFilterClickHandler}>Add</button></span>
+            </div>
+            <div className="row">
+                <ul>{filtersApplied}</ul>
+            </div>
+            <div className="row">
+                <span className="white-text">Total of cards: {totalCards}</span>
+            </div>
+        </div>
+    )
+}
+
+const
+    mapStateToProps = ({ search }) => ({
+        totalCards: search.result.length,
+        appliedFilters: search.filters,
+        fieldOptions: getFieldsToSearch( search.cardsDictionary[ Object.keys(search.cardsDictionary)[4] ] )
+    }),
+    mapDispatchToProps = { searchCards, onFilterAdd: addFilter, onFilterRemove: removeFilter }
+export default connect(mapStateToProps, mapDispatchToProps)(FilterBox)
